requirements.txt
flask
flask-cors
pydantic
numpy
scipy
google-generativeai
pytest
gunicorn
python-dotenv
